feat: add scalable module router for gpt modules
