+++ src/routes/register.ts
@@ -4,7 +4,7 @@ import arcanosRouter from './arcanos.js';
 import arcanosPipelineRouter from './openai-arcanos-pipeline.js';
 import aiEndpointsRouter from './ai-endpoints.js';
 import memoryRouter from './memory.js';
-import tutorRouter from './tutor.js';
+import modulesRouter from './modules.js';
 import workersRouter from './workers.js';
 import heartbeatRouter from './heartbeat.js';
 import orchestrationRouter from './orchestration.js';
@@ -27,7 +27,7 @@ export function registerRoutes(app: Express): void {
   app.use('/', arcanosPipelineRouter);
   app.use('/', aiEndpointsRouter);
   app.use('/', memoryRouter);
-  app.use('/', tutorRouter);
+  app.use('/', modulesRouter);
   app.use('/', workersRouter);
   app.use('/', heartbeatRouter);
   app.use('/', orchestrationRouter);
